requirements/base.txt
coreapi
coreapi-cli
django
django-allauth
django-cors-headers
django-environ
django-rest-auth
djangorestframework
djangorestframework-jwt
gunicorn
psycopg2
